+++ ydb/public/sdk/cpp/CHANGELOG.md
@@ -1,5 +1,7 @@
 * Added automatic retries for unary methods of table and query clients(ExecuteQuery, ExecuteScript, BulkUpsert, ReadRows).
 
+* Implemented native ranges(TRowRange) and iterators over both streaming query results and TResultSet.
+
 ## v3.19.0
 
 * Added metric buffer for batched metric writes to reduce OpenTelemetry hot-path overhead.
@@ -13,7 +15,6 @@
 * Fixed self thread join core dump in IAM credentials provider
 
 * Fixed some data races in Producer (concurrent access to `PartitionsIndex`)
-
 
 * Added interface for export of metrics and spans, supported plugin for OpenTelemetry